+++ distro/config.go
@@ -16,9 +16,11 @@ package distro
 
 import (
 	"crypto/tls"
+	"fmt"
 	"os"
 	"strings"
 
+	"github.com/go-logr/logr"
 	"go.opentelemetry.io/contrib/propagators/aws/xray"
 	"go.opentelemetry.io/contrib/propagators/b3"
 	"go.opentelemetry.io/contrib/propagators/jaeger"
@@ -43,6 +45,9 @@ const (
 	otelExporterOTLPEndpointKey       = "OTEL_EXPORTER_OTLP_ENDPOINT"
 	otelExporterOTLPTracesEndpointKey = "OTEL_EXPORTER_OTLP_TRACES_ENDPOINT"
 
+	// Logging level to set when using the default logger.
+	otelLogLevelKey = "OTEL_LOG_LEVEL"
+
 	// FIXME: support OTEL_SPAN_LINK_COUNT_LIMIT
 	// FIXME: support OTEL_ATTRIBUTE_VALUE_LENGTH_LIMIT
 )
@@ -52,6 +57,7 @@ const (
 	defaultAccessToken   = ""
 	defaultPropagator    = "tracecontext,baggage"
 	defaultTraceExporter = "otlp"
+	defaultLogLevel      = "info"
 
 	defaultOTLPEndpoint   = "localhost:4317"
 	defaultJaegerEndpoint = "http://127.0.0.1:9080/v1/trace"
@@ -67,6 +73,7 @@ type exporterConfig struct {
 
 // config is the configuration used to create and operate an SDK.
 type config struct {
+	Logger     logr.Logger
 	Propagator propagation.TextMapPropagator
 
 	ExportConfig      *exporterConfig
@@ -76,6 +83,7 @@ type config struct {
 // newConfig returns a validated config with Splunk defaults.
 func newConfig(opts ...Option) *config {
 	c := &config{
+		Logger: logger(zapConfig(envOr(otelLogLevelKey, defaultLogLevel))),
 		ExportConfig: &exporterConfig{
 			AccessToken: envOr(accessTokenKey, defaultAccessToken),
 		},
@@ -87,15 +95,15 @@ func newConfig(opts ...Option) *config {
 
 	// Apply default field values if they were not set.
 	if c.Propagator == nil {
-		c.Propagator = loadPropagator(
-			envOr(otelPropagatorsKey, defaultPropagator),
-		)
+		c.setPropagator(envOr(otelPropagatorsKey, defaultPropagator))
 	}
 	if c.TraceExporterFunc == nil {
 		key := envOr(otelTracesExporterKey, defaultTraceExporter)
 		tef, ok := exporters[key]
 		if !ok {
-			// TODO: log invalid exporter value.
+			err := fmt.Errorf("invalid exporter: %q", key)
+			c.Logger.Error(err, "using default trace exporter: otlp")
+
 			tef = exporters[defaultTraceExporter]
 		}
 		c.TraceExporterFunc = tef
@@ -147,13 +155,13 @@ var propagators = map[string]func() propagation.TextMapPropagator{
 	},
 }
 
-func loadPropagator(name string) propagation.TextMapPropagator {
+func (c *config) setPropagator(name string) {
 	var props []propagation.TextMapPropagator
 	for _, part := range strings.Split(name, ",") {
 		factory, ok := propagators[part]
 		if !ok {
 			// Skip invalid data.
-			// TODO: log this.
+			c.Logger.Info("invalid propagator name", "name", part)
 			continue
 		}
 
@@ -161,7 +169,8 @@ func loadPropagator(name string) propagation.TextMapPropagator {
 		if p == nonePropagator {
 			// Make sure the disablement of the global propagator does not get
 			// lost as a composite below.
-			return p
+			c.Propagator = p
+			return
 		}
 		props = append(props, p)
 	}
@@ -169,14 +178,14 @@ func loadPropagator(name string) propagation.TextMapPropagator {
 	switch len(props) {
 	case 0:
 		// Default to "tracecontext,baggage".
-		return propagation.NewCompositeTextMapPropagator(
+		c.Propagator = propagation.NewCompositeTextMapPropagator(
 			propagation.TraceContext{},
 			propagation.Baggage{},
 		)
 	case 1:
-		return props[0]
+		c.Propagator = props[0]
 	default:
-		return propagation.NewCompositeTextMapPropagator(props...)
+		c.Propagator = propagation.NewCompositeTextMapPropagator(props...)
 	}
 }
 
@@ -272,5 +281,22 @@ func WithPropagator(p propagation.TextMapPropagator) Option {
 			p = nonePropagator
 		}
 		c.Propagator = p
+	})
+}
+
+// WithLogger configures the logger used by this distro.
+//
+// The logr.Logger provided should be configured with a verbosity enabled to
+// emit Info logs of the desired level. The following log level to verbosity
+// value are used.
+//   - warning: 0
+//   - info: 1
+//   - debug: 2+
+//
+// By default, a zapr.Logger configured for info logging will be used if this
+// is not provided.
+func WithLogger(l logr.Logger) Option {
+	return optionFunc(func(c *config) {
+		c.Logger = l
 	})
 }